backend/pyproject.toml
[project]
name = "clubresolve"
version = "0.1.0"
description = "Case organization, governance research, and parent advocacy support for BC sports club disputes"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "aiosqlite>=0.20.0",
    # LLM / Agents
    "langgraph>=0.4.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "litellm>=1.50.0",
    # Vector store
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    # Encryption
    "cryptography>=43.0.0",
    # Document parsing
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    # PDF export
    "weasyprint>=63.0",
    # Utils
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
